requirements.txt
google-generativeai>=0.5
google-adk>=0.4
wikipedia-api>=0.6
requests>=2.32
jinja2>=3.1
python-dotenv>=1.0
certifi>=2024.0
